refactor(tests): consolidate graphify Module — 7 files → 1
Closes #3761
